+++ src/main/java/com/example/basicapp/PersonRowMapper.java
@@ -10,7 +10,7 @@ public class PersonRowMapper implements RowMapper<PersonDAO> {
     @Override
     public PersonDAO mapRow(ResultSet rs, int row) throws SQLException {
         var person = new PersonDAO();
-        person.setPersonId(rs.getInt("person_id"));
+        person.setPersonId(rs.getLong("person_id"));
         person.setName(rs.getString("name"));
         person.setAge(rs.getInt("age"));
         return person;
